pyproject.toml
[tool.poetry]
name = "jet"
version = "0.1.0"
description = ""
authors = ["Volltin <volltin@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
openai = "^0.28.1"
langchain = "^0.0.330"
python-dotenv = "^1.0.0"
gradio = "^4.1.1"
click = "^8.1.7"
sqlitedict = "^2.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
pre-commit = "^3.4.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
py_version = 310
line_length = 88
length_sort = true
